Revert "템플릿 페이지 QA 반영 (#100)"
This reverts commit cbb16fccdb34e9c2193afebf87ecd548780a7cd9.
